music_service/requirements.txt
Django==3.0.5
djangorestframework==3.10.0
psycopg2-binary==2.8.5
dj-rest-auth==1.0.4
django-allauth==0.41.0
django-cors-headers==3.5.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
Pillow==7.1.1
coverage==5.2.1
flake8===3.7.9
black==19.10b0
isort==4.3.21
